+++ frontend/awx/settings/Settings.tsx
@@ -1,13 +1,15 @@
 import { Fragment, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { PageHeader } from '../../../framework';
-import { RouteObj } from '../../common/Routes';
+import { AwxRoute } from '../AwxRoutes';
+import { useGetAwxUrl } from '../useAwxNavigate';
 
 export default function Settings() {
   const { t } = useTranslation();
+  const getAwxUrl = useGetAwxUrl();
   const breadcrumbs = useMemo(
-    () => [{ label: 'Dashboard', to: RouteObj.Dashboard }, { label: 'Settings' }],
-    []
+    () => [{ label: 'Dashboard', to: getAwxUrl(AwxRoute.Dashboard) }, { label: 'Settings' }],
+    [getAwxUrl]
   );
   return (
     <Fragment>
