+++ frontend/awx/views/schedules/ScheduleForm.tsx
@@ -8,9 +8,11 @@ import { dateToInputDateTime } from '../../../../framework/utils/dateTimeHelpers
 import { RouteObj } from '../../../common/Routes';
 import { postRequest } from '../../../common/crud/Data';
 import { useGet } from '../../../common/crud/useGet';
+import { AwxRoute } from '../../AwxRoutes';
 import { AwxError } from '../../common/AwxError';
 import { getAddedAndRemoved } from '../../common/util/getAddedAndRemoved';
 import { ScheduleFormFields } from '../../interfaces/ScheduleFormFields';
+import { useGetAwxUrl } from '../../useAwxNavigate';
 import { getAwxError } from '../../useAwxView';
 import { ScheduleInputs } from './components/ScheduleInputs';
 import { buildScheduleContainer } from './hooks/scheduleHelpers';
@@ -24,6 +26,7 @@ const routes: { [key: string]: string } = {
 
 export function CreateSchedule() {
   const { t } = useTranslation();
+  const getAwxUrl = useGetAwxUrl();
   const now = DateTime.now();
   const [error, setError] = useState<Error | null>(null);
   const closestQuarterHour: DateTime = DateTime.fromMillis(
@@ -197,7 +200,7 @@ export function CreateSchedule() {
       <PageHeader
         title={t('Create Schedule')}
         breadcrumbs={[
-          { label: t('Schedules'), to: RouteObj.Schedules },
+          { label: t('Schedules'), to: getAwxUrl(AwxRoute.Schedules) },
           { label: t('Create Schedule') },
         ]}
       />
