+++ frontend/awx/views/schedules/SchedulePage/SchedulePage.tsx
@@ -1,4 +1,5 @@
 import { DropdownPosition } from '@patternfly/react-core';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { PageActions, PageHeader, PageLayout } from '../../../../../framework';
@@ -6,8 +7,10 @@ import { LoadingPage } from '../../../../../framework/components/LoadingPage';
 import { PageBackTab, RoutedTab, RoutedTabs } from '../../../../common/RoutedTabs';
 import { RouteObj } from '../../../../common/Routes';
 import { useGetItem } from '../../../../common/crud/useGet';
+import { AwxRoute } from '../../../AwxRoutes';
 import { AwxError } from '../../../common/AwxError';
 import { Schedule } from '../../../interfaces/Schedule';
+import { useGetAwxUrl } from '../../../useAwxNavigate';
 import {
   resourceSchedulePageRoutes,
   scheduleDetailRoutes,
@@ -17,7 +20,6 @@ import {
 import { useSchedulesActions } from '../hooks/useSchedulesActions';
 import { ScheduleDetails } from './ScheduleDetails';
 import { ScheduleRules } from './ScheduleRules';
-import { useMemo } from 'react';
 
 const rulesListRoutes: { [key: string]: string } = {
   inventory: RouteObj.InventorySourceScheduleRules,
@@ -28,6 +30,7 @@ const rulesListRoutes: { [key: string]: string } = {
 
 export function SchedulePage() {
   const { t } = useTranslation();
+  const getAwxUrl = useGetAwxUrl();
   const location = useLocation();
   const params = useParams<{ id: string; source_id?: string; schedule_id: string }>();
   const {
@@ -42,10 +45,10 @@ export function SchedulePage() {
   );
 
   const itemActions = useSchedulesActions({
-    onScheduleToggleorDeleteCompleted: () => navigate(RouteObj.Schedules),
+    onScheduleToggleorDeleteCompleted: () => navigate(getAwxUrl(AwxRoute.Schedules)),
   });
   const generateBackToSchedulesUrl: string = useMemo(() => {
-    if (!resource_type || !schedule) return RouteObj.Schedules;
+    if (!resource_type || !schedule) return getAwxUrl(AwxRoute.Schedules);
     if (resource_type === 'inventory' && schedule?.summary_fields.inventory) {
       return RouteObj.InventorySourceSchedules.replace(':inventory_type', resource_type as string)
         .replace(':id', schedule.summary_fields.inventory.id.toString())
@@ -55,7 +58,7 @@ export function SchedulePage() {
       ':id',
       schedule.summary_fields?.unified_job_template.id.toString()
     );
-  }, [resource_type, schedule]);
+  }, [getAwxUrl, resource_type, schedule]);
   if (error) return <AwxError error={error} handleRefresh={refresh} />;
   if (!schedule) return <LoadingPage breadcrumbs tabs />;
   return (
@@ -84,13 +87,13 @@ export function SchedulePage() {
         />
         <RoutedTab
           label={t('Details')}
-          url={resource_type ? scheduleDetailRoutes[resource_type] : RouteObj.Schedules}
+          url={resource_type ? scheduleDetailRoutes[resource_type] : getAwxUrl(AwxRoute.Schedules)}
         >
           <ScheduleDetails schedule={schedule} />
         </RoutedTab>
         <RoutedTab
           label={t(`Rules`)}
-          url={resource_type ? rulesListRoutes[resource_type] : RouteObj.Schedules}
+          url={resource_type ? rulesListRoutes[resource_type] : getAwxUrl(AwxRoute.Schedules)}
         >
           <ScheduleRules rrule={schedule.rrule} />
         </RoutedTab>
