+++ frontend/common/LoginModal.tsx
@@ -6,6 +6,8 @@ import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { PageForm, PageFormSubmitHandler, usePageDialog } from '../../framework';
 import { PageFormTextInput } from '../../framework/PageForm/Inputs/PageFormTextInput';
+import { AwxRoute } from '../awx/AwxRoutes';
+import { useAwxNavigate } from '../awx/useAwxNavigate';
 import { hubAPI } from '../hub/api/utils';
 import { RouteObj } from './Routes';
 import { AuthOptions, SocialAuthLogin } from './SocialAuthLogin';
@@ -100,6 +102,7 @@ function LoginForm(props: LoginFormProps) {
   const { authOptions } = props;
   const { t } = useTranslation();
   const navigate = useNavigate();
+  const awxNavigate = useAwxNavigate();
 
   useInvalidateCacheOnUnmount();
 
@@ -173,7 +176,7 @@ function LoginForm(props: LoginFormProps) {
         }
         switch (process.env.UI_MODE) {
           case 'AWX':
-            navigate(RouteObj.Dashboard);
+            awxNavigate(AwxRoute.Dashboard);
             break;
           case 'EDA':
             navigate(RouteObj.EdaDashboard);
@@ -193,7 +196,7 @@ function LoginForm(props: LoginFormProps) {
         }
       }
     },
-    [navigate, props, t]
+    [awxNavigate, navigate, props, t]
   );
 
   return (
